pyproject.toml
[tool.poetry]
name = "streamlit-ghg-consulting"
version = "0.1.0"
description = ""
authors = ["Alexander Orvik <alexao1111@gmail.com>"]
readme = "README.md"
packages = [{include = "streamlit_ghg_consulting"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.29.0"
replicate = "^0.22.0"
pinecone-client = "^2.2.4"
transformers = "^4.36.2"
torch = "^2.1.2"
llama-index = "^0.9.25.post1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
llmsherpa = "^0.1.3"
pypdf = "^3.17.4"
watchdog = "^3.0.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
